faststore-poc(SFS-1709): test custom server
